--- client/src/app/ape/results/[id]/page.js
@@ -1,52 +1,0 @@
-"use client";
-
-import { useEffect, useState } from "react";
-import { useParams } from "next/navigation";
-import Status from "../../status";
-export default function ResultsPage() {
-  const { id } = useParams();
-  const [status, setStatus] = useState(null);
-  const [seerData, setSeerData] = useState(null);
-
-  useEffect(() => {
-    if (!id) return;
-
-    const interval = setInterval(async () => {
-      try {
-        const res = await fetch(`/api/status/${id}`);
-        const data = await res.json();
-
-        setStatus(data.status);
-        setSeerData(data.seerData || null);
-
-        // stop polling when job is done or failed
-        if (["COMPLETED", "FAILED"].includes(data.status?.status)) {
-          clearInterval(interval);
-        }
-      } catch (err) {
-        console.error("Failed to fetch job status:", err);
-        clearInterval(interval);
-      }
-    }, 3000);
-
-    return () => clearInterval(interval);
-  }, [id]);
-
-  return (
-    <div className="container py-4">
-      <h3>Job Status</h3>
-      <div>
-        <strong>Job ID:</strong> {id}
-      </div>
-
-      <Status status={status} seerData={seerData} />
-
-      {status?.status === "COMPLETED" && (
-        <div className="mt-4">
-          <h5>Results:</h5>
-          <pre>{JSON.stringify(seerData, null, 2)}</pre>
-        </div>
-      )}
-    </div>
-  );
-}
